solution_pseudo-random/requirements.txt
ecdsa
pyjwt
jwcrypto
requests
